+++ mobile/android/android-components/components/browser/engine-gecko/src/main/java/mozilla/components/browser/engine/gecko/GeckoEngine.kt
@@ -381,6 +381,7 @@ class GeckoEngine(
                         NativePermissionPromptResponse(
                             data.isPermissionsGranted,
                             data.isPrivateModeGranted,
+                            false, // isTechnicalAndInteractionDataGranted
                         ),
                     )
                 }