+++ mobile/android/geckoview/src/androidTest/java/org/mozilla/geckoview/test/WebExtensionTest.kt
@@ -614,9 +614,27 @@ class WebExtensionTest : BaseSessionTest() {
             ),
         )
 
+        sessionRule.delegateDuringNextWait(object : WebExtensionController.PromptDelegate {
+            @AssertCalled
+            override fun onInstallPromptRequest(
+                extension: WebExtension,
+                permissions: Array<String>,
+                origins: Array<String>,
+                dataCollectionPermissions: Array<String>,
+            ): GeckoResult<PermissionPromptResponse>? {
+                return GeckoResult.fromValue(
+                    PermissionPromptResponse(
+                        true, // isPermissionsGranted
+                        false, // isPrivateModeGranted
+                        false, // isTechnicalAndInteractionDataGranted
+                    ),
+                )
+            }
+        })
+
         val extension = sessionRule.waitForResult(
-            controller.ensureBuiltIn(
-                "resource://android/assets/web_extensions/data-collection/",
+            controller.install(
+                "resource://android/assets/web_extensions/data-collection-unsigned.xpi",
                 "data-collection@test.mozilla.org",
             ),
         )
@@ -785,6 +803,7 @@ class WebExtensionTest : BaseSessionTest() {
                     PermissionPromptResponse(
                         true, // isPermissionsGranted
                         true, // isPrivateModeGranted
+                        false, // isTechnicalAndInteractionDataGranted
                     ),
                 )
             }
@@ -898,6 +917,7 @@ class WebExtensionTest : BaseSessionTest() {
                     PermissionPromptResponse(
                         true, // isPermissionsGranted
                         false, // isPrivateModeGranted
+                        false, // isTechnicalAndInteractionDataGranted
                     ),
                 )
             }
@@ -973,6 +993,7 @@ class WebExtensionTest : BaseSessionTest() {
                     PermissionPromptResponse(
                         true, // isPermissionsGranted
                         false, // isPrivateModeGranted
+                        false, // isTechnicalAndInteractionDataGranted
                     ),
                 )
             }
@@ -1051,6 +1072,7 @@ class WebExtensionTest : BaseSessionTest() {
                     PermissionPromptResponse(
                         true, // isPermissionsGranted
                         true, // isPrivateModeGranted
+                        false, // isTechnicalAndInteractionDataGranted
                     ),
                 )
             }
@@ -1111,6 +1133,7 @@ class WebExtensionTest : BaseSessionTest() {
                     PermissionPromptResponse(
                         true, // isPermissionsGranted
                         false, // isPrivateModeGranted
+                        false, // isTechnicalAndInteractionDataGranted
                     ),
                 )
             }
@@ -1200,6 +1223,7 @@ class WebExtensionTest : BaseSessionTest() {
                     PermissionPromptResponse(
                         true, // isPermissionsGranted
                         false, // isPrivateModeGranted
+                        false, // isTechnicalAndInteractionDataGranted
                     ),
                 )
             }
@@ -1281,6 +1305,7 @@ class WebExtensionTest : BaseSessionTest() {
                     PermissionPromptResponse(
                         true, // isPermissionsGranted
                         false, // isPrivateModeGranted
+                        false, // isTechnicalAndInteractionDataGranted
                     ),
                 )
             }
@@ -1342,6 +1367,7 @@ class WebExtensionTest : BaseSessionTest() {
                     PermissionPromptResponse(
                         true, // isPermissionsGranted
                         false, // isPrivateModeGranted
+                        false, // isTechnicalAndInteractionDataGranted
                     ),
                 )
             }
@@ -1409,6 +1435,7 @@ class WebExtensionTest : BaseSessionTest() {
                     PermissionPromptResponse(
                         true, // isPermissionsGranted
                         false, // isPrivateModeGranted
+                        false, // isTechnicalAndInteractionDataGranted
                     ),
                 )
             }
@@ -1513,6 +1540,7 @@ class WebExtensionTest : BaseSessionTest() {
                     PermissionPromptResponse(
                         true, // isPermissionsGranted
                         false, // isPrivateModeGranted
+                        false, // isTechnicalAndInteractionDataGranted
                     ),
                 )
             }
@@ -1583,6 +1611,7 @@ class WebExtensionTest : BaseSessionTest() {
                     PermissionPromptResponse(
                         false, // isPermissionsGranted
                         false, // isPrivateModeGranted
+                        false, // isTechnicalAndInteractionDataGranted
                     ),
                 )
             }
@@ -2199,6 +2228,7 @@ class WebExtensionTest : BaseSessionTest() {
                     PermissionPromptResponse(
                         true, // isPermissionsGranted
                         false, // isPrivateModeGranted
+                        false, // isTechnicalAndInteractionDataGranted
                     ),
                 )
             }
@@ -2283,6 +2313,7 @@ class WebExtensionTest : BaseSessionTest() {
                     PermissionPromptResponse(
                         true, // isPermissionsGranted
                         false, // isPrivateModeGranted
+                        false, // isTechnicalAndInteractionDataGranted
                     ),
                 )
             }
@@ -2303,6 +2334,7 @@ class WebExtensionTest : BaseSessionTest() {
                     PermissionPromptResponse(
                         true, // isPermissionsGranted
                         false, // isPrivateModeGranted
+                        false, // isTechnicalAndInteractionDataGranted
                     ),
                 )
             }
@@ -3226,6 +3258,7 @@ class WebExtensionTest : BaseSessionTest() {
                     PermissionPromptResponse(
                         true, // isPermissionsGranted
                         false, // isPrivateModeGranted
+                        false, // isTechnicalAndInteractionDataGranted
                     ),
                 )
             }
@@ -3293,6 +3326,7 @@ class WebExtensionTest : BaseSessionTest() {
                     PermissionPromptResponse(
                         true, // isPermissionsGranted
                         false, // isPrivateModeGranted
+                        false, // isTechnicalAndInteractionDataGranted
                     ),
                 )
             }
@@ -3340,6 +3374,7 @@ class WebExtensionTest : BaseSessionTest() {
                     PermissionPromptResponse(
                         true, // isPermissionsGranted
                         false, // isPrivateModeGranted
+                        false, // isTechnicalAndInteractionDataGranted
                     ),
                 )
             }
@@ -3394,6 +3429,7 @@ class WebExtensionTest : BaseSessionTest() {
                     PermissionPromptResponse(
                         true, // isPermissionsGranted
                         false, // isPrivateModeGranted
+                        false, // isTechnicalAndInteractionDataGranted
                     ),
                 )
             }
@@ -3473,6 +3509,7 @@ class WebExtensionTest : BaseSessionTest() {
                     PermissionPromptResponse(
                         true, // isPermissionsGranted
                         false, // isPrivateModeGranted
+                        false, // isTechnicalAndInteractionDataGranted
                     ),
                 )
             }
@@ -3536,6 +3573,7 @@ class WebExtensionTest : BaseSessionTest() {
                     PermissionPromptResponse(
                         true, // isPermissionsGranted
                         false, // isPrivateModeGranted
+                        false, // isTechnicalAndInteractionDataGranted
                     ),
                 )
             }
@@ -3616,6 +3654,7 @@ class WebExtensionTest : BaseSessionTest() {
                     PermissionPromptResponse(
                         true, // isPermissionsGranted
                         false, // isPrivateModeGranted
+                        false, // isTechnicalAndInteractionDataGranted
                     ),
                 )
             }
@@ -3664,6 +3703,7 @@ class WebExtensionTest : BaseSessionTest() {
                     PermissionPromptResponse(
                         true, // isPermissionsGranted
                         false, // isPrivateModeGranted
+                        false, // isTechnicalAndInteractionDataGranted
                     ),
                 )
             }
@@ -3731,6 +3771,7 @@ class WebExtensionTest : BaseSessionTest() {
                     PermissionPromptResponse(
                         true, // isPermissionsGranted
                         false, // isPrivateModeGranted
+                        false, // isTechnicalAndInteractionDataGranted
                     ),
                 )
             }
@@ -3896,6 +3937,7 @@ class WebExtensionTest : BaseSessionTest() {
                     PermissionPromptResponse(
                         true, // isPermissionsGranted
                         false, // isPrivateModeGranted
+                        false, // isTechnicalAndInteractionDataGranted
                     ),
                 )
             }
@@ -3981,6 +4023,7 @@ class WebExtensionTest : BaseSessionTest() {
                     PermissionPromptResponse(
                         true, // isPermissionsGranted
                         false, // isPrivateModeGranted
+                        false, // isTechnicalAndInteractionDataGranted
                     ),
                 )
             }
@@ -4284,6 +4327,7 @@ class WebExtensionTest : BaseSessionTest() {
                     PermissionPromptResponse(
                         true, // isPermissionsGranted
                         false, // isPrivateModeGranted
+                        false, // isTechnicalAndInteractionDataGranted
                     ),
                 )
             }
@@ -4426,6 +4470,7 @@ class WebExtensionTest : BaseSessionTest() {
                     PermissionPromptResponse(
                         true, // isPermissionsGranted
                         false, // isPrivateModeGranted
+                        false, // isTechnicalAndInteractionDataGranted
                     ),
                 )
             }
@@ -4454,5 +4499,59 @@ class WebExtensionTest : BaseSessionTest() {
         mainSession.loadUri("about:crashextensions")
 
         sessionRule.waitForResult(controller.uninstall(borderify))
+    }
+
+    @Test
+    fun installWebExtensionWithTechnicalAndInteractionData() {
+        sessionRule.setPrefsUntilTestEnd(
+            mapOf(
+                "xpinstall.signatures.required" to false,
+                "extensions.dataCollectionPermissions.enabled" to true,
+            ),
+        )
+
+        sessionRule.delegateDuringNextWait(object : WebExtensionController.PromptDelegate {
+            @AssertCalled
+            override fun onInstallPromptRequest(
+                extension: WebExtension,
+                permissions: Array<String>,
+                origins: Array<String>,
+                dataCollectionPermissions: Array<String>,
+            ): GeckoResult<PermissionPromptResponse>? {
+                return GeckoResult.fromValue(
+                    // We grant the `technicalAndInteraction` data collection
+                    // permission below.
+                    PermissionPromptResponse(
+                        true, // isPermissionsGranted
+                        false, // isPrivateModeGranted
+                        true, // isTechnicalAndInteractionDataGranted
+                    ),
+                )
+            }
+        })
+
+        val extension = sessionRule.waitForResult(
+            controller.install(
+                "resource://android/assets/web_extensions/data-collection-unsigned.xpi",
+                "data-collection@test.mozilla.org",
+            ),
+        )
+        assertEquals("data-collection@test.mozilla.org", extension.id)
+
+        var optionalDataCollectionPermissions = extension.metaData.optionalDataCollectionPermissions
+        assertArrayEquals(
+            "optionalDataCollectionPermissions has the expected permissions",
+            arrayOf("technicalAndInteraction", "locationInfo"),
+            optionalDataCollectionPermissions,
+        )
+
+        var grantedOptionalDataCollectionPermissions = extension.metaData.grantedOptionalDataCollectionPermissions
+        assertArrayEquals(
+            "grantedOptionalDataCollectionPermissions has the expected permissions",
+            arrayOf("technicalAndInteraction"),
+            grantedOptionalDataCollectionPermissions,
+        )
+
+        sessionRule.waitForResult(controller.uninstall(extension))
     }
 }